Some minor tweaks to show new metrics
