+++ torchci/components/benchmark/llms/SummaryPanel.tsx
@@ -2,7 +2,6 @@ import { Grid } from "@mui/material";
 import { GridCellParams, GridRenderCellParams } from "@mui/x-data-grid";
 import {
   BranchAndCommitPerfData,
-  LLMsBenchmarkData,
   METRIC_DISPLAY_HEADERS,
   RELATIVE_THRESHOLD,
 } from "components/benchmark/llms/common";
@@ -10,6 +9,7 @@ import styles from "components/metrics.module.css";
 import { TablePanelWithData } from "components/metrics/panels/TablePanel";
 import { Granularity } from "components/metrics/panels/TimeSeriesPanel";
 import dayjs from "dayjs";
+import { combineLeftAndRight } from "lib/benchmark/llmUtils";
 
 const ROW_GAP = 100;
 const ROW_HEIGHT = 38;
@@ -40,108 +40,7 @@ export function SummaryPanel({
   const rCommit = rPerfData.commit;
   const rData = rPerfData.data;
 
-  const dataGroupedByModel: { [k: string]: any } = {};
-  rData.forEach((record: LLMsBenchmarkData) => {
-    const name = record.name;
-    const dtype = record.dtype;
-    const device = record.device;
-    const metric = record.metric;
-
-    if (!(name in dataGroupedByModel)) {
-      dataGroupedByModel[name] = {};
-    }
-
-    if (!(dtype in dataGroupedByModel[name])) {
-      dataGroupedByModel[name][dtype] = {};
-    }
-
-    if (!(device in dataGroupedByModel[name][dtype])) {
-      dataGroupedByModel[name][dtype][device] = {};
-    }
-
-    dataGroupedByModel[name][dtype][device][metric] = {
-      r: record,
-    };
-  });
-
-  // Combine with left (base) data
-  if (lCommit !== rCommit && lData !== undefined) {
-    lData.forEach((record: LLMsBenchmarkData) => {
-      const name = record.name;
-      const dtype = record.dtype;
-      const device = record.device;
-      const metric = record.metric;
-
-      if (!(name in dataGroupedByModel)) {
-        dataGroupedByModel[name] = {};
-      }
-
-      if (!(dtype in dataGroupedByModel[name])) {
-        dataGroupedByModel[name][dtype] = {};
-      }
-
-      if (!(device in dataGroupedByModel[name][dtype])) {
-        dataGroupedByModel[name][dtype][device] = {};
-      }
-
-      if (!(metric in dataGroupedByModel[name][dtype][device])) {
-        dataGroupedByModel[name][dtype][device][metric] = {};
-      }
-
-      dataGroupedByModel[name][dtype][device][metric]["l"] = record;
-    });
-  }
-
-  // Transform the data into a displayable format
-  const data: { [k: string]: any }[] = [];
-  Object.keys(dataGroupedByModel).forEach((name: string) => {
-    Object.keys(dataGroupedByModel[name]).forEach((dtype: string) => {
-      Object.keys(dataGroupedByModel[name][dtype]).forEach((device: string) => {
-        const row: { [k: string]: any } = {
-          // Keep the name as as the row ID as DataGrid requires it
-          name: `${name} (${dtype} / ${device})`,
-        };
-
-        for (const metric in dataGroupedByModel[name][dtype][device]) {
-          const record = dataGroupedByModel[name][dtype][device][metric];
-          const hasL = "l" in record;
-          const hasR = "r" in record;
-
-          row["metadata"] = {
-            name: name,
-            dtype: dtype,
-            device: device,
-            l: hasL ? record["l"]["job_id"] : undefined,
-            r: hasR ? record["r"]["job_id"] : undefined,
-          };
-
-          row[metric] = {
-            l: hasL
-              ? {
-                  actual: record["l"].actual,
-                  target: record["l"].target,
-                }
-              : {
-                  actual: 0,
-                  target: 0,
-                },
-            r: hasR
-              ? {
-                  actual: record["r"].actual,
-                  target: record["r"].target,
-                }
-              : {
-                  actual: 0,
-                  target: 0,
-                },
-          };
-        }
-
-        data.push(row);
-      });
-    });
-  });
-
+  const data = combineLeftAndRight(lPerfData, rPerfData);
   return (
     <Grid container spacing={2} style={{ height: "100%" }}>
       <Grid item xs={12} lg={12} height={data.length * ROW_HEIGHT + ROW_GAP}>
@@ -179,7 +78,10 @@ export function SummaryPanel({
                 )}&dtypeName=${encodeURIComponent(
                   dtype
                 )}&deviceName=${encodeURIComponent(device)}`;
-                const isNewModel = params.value.l === undefined ? "✅" : "";
+                const isNewModel =
+                  params.value.l === undefined && lCommit !== rCommit
+                    ? "(NEW!) "
+                    : "";
                 const isModelStopRunning =
                   params.value.r === undefined ? "❌" : "";
 
@@ -260,11 +162,12 @@ export function SummaryPanel({
                   const rPercent = target
                     ? `(${Number((r * 100) / target).toFixed(0)}%)`
                     : "";
+                  const isNewModel = l === 0 ? "(NEW!)" : "";
 
-                  if (lCommit === rCommit || l === r || v.l === 0) {
+                  if (lCommit === rCommit || l === r) {
                     return `${r} ${rPercent} [target = ${target}]`;
                   } else {
-                    return `${l} ${lPercent} → ${r} ${rPercent} [target = ${target}]`;
+                    return `${l} ${lPercent} → ${r} ${rPercent} [target = ${target}] ${isNewModel} `;
                   }
                 },
               };
