+++ torchci/lib/benchmark/llmUtils.ts
@@ -1,7 +1,9 @@
 import {
+  BranchAndCommitPerfData,
   DEFAULT_DEVICE_NAME,
   DEFAULT_DTYPE_NAME,
   DEFAULT_MODEL_NAME,
+  LLMsBenchmarkData,
 } from "components/benchmark/llms/common";
 import { fetcher } from "lib/GeneralUtils";
 import { RocksetParam } from "lib/rockset";
@@ -66,4 +68,131 @@ export function useBenchmark(
   return useSWR(lUrl, fetcher, {
     refreshInterval: 60 * 60 * 1000, // refresh every hour
   });
+}
+
+export function combineLeftAndRight(
+  lPerfData: BranchAndCommitPerfData,
+  rPerfData: BranchAndCommitPerfData
+): { [k: string]: any }[] {
+  // The left (base commit)
+  const lBranch = lPerfData.branch;
+  const lCommit = lPerfData.commit;
+  const lData = lPerfData.data;
+  // and the right (new commit)
+  const rBranch = rPerfData.branch;
+  const rCommit = rPerfData.commit;
+  const rData = rPerfData.data;
+
+  const dataGroupedByModel: { [k: string]: any } = {};
+  rData.forEach((record: LLMsBenchmarkData) => {
+    const name = record.name;
+    const dtype = record.dtype;
+    const device = record.device;
+    const metric = record.metric;
+
+    if (!(name in dataGroupedByModel)) {
+      dataGroupedByModel[name] = {};
+    }
+
+    if (!(dtype in dataGroupedByModel[name])) {
+      dataGroupedByModel[name][dtype] = {};
+    }
+
+    if (!(device in dataGroupedByModel[name][dtype])) {
+      dataGroupedByModel[name][dtype][device] = {};
+    }
+
+    dataGroupedByModel[name][dtype][device][metric] = {
+      r: record,
+    };
+  });
+
+  // Combine with left (base) data
+  if (lCommit !== rCommit && lData !== undefined) {
+    lData.forEach((record: LLMsBenchmarkData) => {
+      const name = record.name;
+      const dtype = record.dtype;
+      const device = record.device;
+      const metric = record.metric;
+
+      if (!(name in dataGroupedByModel)) {
+        dataGroupedByModel[name] = {};
+      }
+
+      if (!(dtype in dataGroupedByModel[name])) {
+        dataGroupedByModel[name][dtype] = {};
+      }
+
+      if (!(device in dataGroupedByModel[name][dtype])) {
+        dataGroupedByModel[name][dtype][device] = {};
+      }
+
+      if (!(metric in dataGroupedByModel[name][dtype][device])) {
+        dataGroupedByModel[name][dtype][device][metric] = {};
+      }
+
+      dataGroupedByModel[name][dtype][device][metric]["l"] = record;
+    });
+  }
+
+  // Transform the data into a displayable format
+  const data: { [k: string]: any }[] = [];
+  Object.keys(dataGroupedByModel).forEach((name: string) => {
+    Object.keys(dataGroupedByModel[name]).forEach((dtype: string) => {
+      Object.keys(dataGroupedByModel[name][dtype]).forEach((device: string) => {
+        const row: { [k: string]: any } = {
+          // Keep the name as as the row ID as DataGrid requires it
+          name: `${name} (${dtype} / ${device})`,
+        };
+
+        for (const metric in dataGroupedByModel[name][dtype][device]) {
+          const record = dataGroupedByModel[name][dtype][device][metric];
+          const hasL = "l" in record;
+          const hasR = "r" in record;
+
+          if (!("metadata" in row)) {
+            row["metadata"] = {
+              name: name,
+              dtype: dtype,
+              device: device,
+              l: hasL ? record["l"]["job_id"] : undefined,
+              r: hasR ? record["r"]["job_id"] : undefined,
+            };
+          } else {
+            row["metadata"]["l"] =
+              row["metadata"]["l"] ??
+              (hasL ? record["l"]["job_id"] : undefined);
+            row["metadata"]["r"] =
+              row["metadata"]["r"] ??
+              (hasR ? record["r"]["job_id"] : undefined);
+          }
+
+          row[metric] = {
+            l: hasL
+              ? {
+                  actual: record["l"].actual,
+                  target: record["l"].target,
+                }
+              : {
+                  actual: 0,
+                  target: 0,
+                },
+            r: hasR
+              ? {
+                  actual: record["r"].actual,
+                  target: record["r"].target,
+                }
+              : {
+                  actual: 0,
+                  target: 0,
+                },
+          };
+        }
+
+        data.push(row);
+      });
+    });
+  });
+
+  return data;
 }
